Do not add empty attributes to the document
Currently any empty attributes including blank spaces are added to the document. However I think it is not needed to add empty fields to the document. So I have added some changes. filterAttribute property can define whether to remove the empty values or not. And I used by overriding the save method. 
Maybe there is a better way to do it.