Fix inserted _id on insert of EmbedsMany
When an object is associated to parent in `associateNew`, a Mongo ID is created and added to the model regardless of the key set on the model. This commit changes the behaviour to only add the _id in instances where the primary key is set to be `_id`. This _may_ be a breaking change if code is depending on this behaviour; but the existing behaviour is undocumented.